pyproject.toml
[tool.poetry]
name = "porte"
version = "0.1.0"
description = ""
authors = ["Vincent Bisserié <vbisserie@protonmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
pyudev = "^0.24.0"
"RPi.GPIO" = "^0.7.1"
requests = "^2.28.2"
sentry-sdk = "^1.15.0"
dotmap = "^1.3.30"
ruamel-yaml = "^0.17.21"
tinydb = "^4.7.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
